pyproject.toml
[project]
name = "styletts2"
version = "0.1.0"
description = "StyleTTS 2: Towards Human-Level Text-to-Speech through Style Diffusion and Adversarial Training"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "SoundFile",
    "torchaudio",
    "munch",
    "torch>=2.0.0",
    "pydub",
    "pyyaml",
    "librosa",
    "nltk",
    "matplotlib",
    "accelerate",
    "transformers",
    "einops",
    "einops-exts",
    "tqdm",
    "typing",
    "typing-extensions",
    "monotonic-align",
    "pandas>=2.2.3",
    "tensorboard>=2.19.0",
    "mecab-python3==1.0.9",
    "fugashi[unidic-lite]==1.3.0",
    "misaki[en,ja]>=0.9.3",
    "pyopenjtalk-plus",
    "psutil>=7.0.0",
    "jupyter>=1.1.1",
    "ipykernel>=6.29.5",
    "torchvision>=0.21.0",
    "timm>=1.0.15",
    "gdown>=5.2.0",
    "ipdb>=0.13.13",
]

[tool.setuptools]
py-modules = ["models", "meldataset", "losses", "optimizers", "text_utils", "utils"]

[tool.setuptools.packages.find]
exclude = ["Data*", "Demo*", "Colab*", "Configs*", "Utils*", "Modules*"]

[tool.uv]
package = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu124" }]
torchaudio = [{ index = "pytorch-cu124" }]
monotonic-align = { git = "https://github.com/resemble-ai/monotonic_align.git" }

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

